split out settings (stored in db) to separate file
