+++ cmd/relay/main.go
@@ -4,11 +4,9 @@ import (
 	"context"
 	"crypto/rand"
 	"encoding/base64"
-	"errors"
 	"fmt"
 	"github.com/bluesky-social/indigo/atproto/identity"
 	"github.com/bluesky-social/indigo/cmd/relay/events/diskpersist"
-	"gorm.io/gorm"
 	"io"
 	"log/slog"
 	_ "net/http/pprof"
@@ -16,7 +14,6 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"syscall"
 	"time"
@@ -425,54 +422,4 @@ func makePdsClientSetup(ratelimitBypass string) func(c *xrpc.Client) {
 			c.Client.Timeout = time.Minute * 1
 		}
 	}
-}
-
-// RelaySetting is a gorm model
-type RelaySetting struct {
-	Name  string `gorm:"primarykey"`
-	Value string
-}
-
-func getRelaySetting(db *gorm.DB, name string) (value string, found bool, err error) {
-	var setting RelaySetting
-	dbResult := db.First(&setting, "name = ?", name)
-	if errors.Is(dbResult.Error, gorm.ErrRecordNotFound) {
-		return "", false, nil
-	}
-	if dbResult.Error != nil {
-		return "", false, dbResult.Error
-	}
-	return setting.Value, true, nil
-}
-
-func setRelaySetting(db *gorm.DB, name string, value string) error {
-	return db.Transaction(func(tx *gorm.DB) error {
-		var setting RelaySetting
-		found := tx.First(&setting, "name = ?", name)
-		if errors.Is(found.Error, gorm.ErrRecordNotFound) {
-			// ok! create it
-			setting.Name = name
-			setting.Value = value
-			return tx.Create(&setting).Error
-		} else if found.Error != nil {
-			return found.Error
-		}
-		setting.Value = value
-		return tx.Save(&setting).Error
-	})
-}
-
-func getRelaySettingBool(db *gorm.DB, name string) (value bool, found bool, err error) {
-	strval, found, err := getRelaySetting(db, name)
-	if err != nil || !found {
-		return false, found, err
-	}
-	value, err = strconv.ParseBool(strval)
-	if err != nil {
-		return false, false, err
-	}
-	return value, true, nil
-}
-func setRelaySettingBool(db *gorm.DB, name string, value bool) error {
-	return setRelaySetting(db, name, strconv.FormatBool(value))
 }
